backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
google-generativeai
sqlalchemy
pydantic
python-dotenv
fpdf2
python-docx
